fix: Correct (D.6) key issue and update argument types
- Combine functions to perform state merklization (5.8)
- Fix a bug that excluded the first bit of the key
